+++ src/video_core/regs_framebuffer.h
@@ -193,7 +193,7 @@ struct FramebufferRegs {
         case ColorFormat::RGBA4:
             return 2;
         default:
-            LOG_CRITICAL(HW_GPU, "Unknown color format %u", format);
+            LOG_CRITICAL(HW_GPU, "Unknown color format %u", static_cast<int>(format));
             UNIMPLEMENTED();
         }
     }
@@ -256,7 +256,7 @@ struct FramebufferRegs {
         case DepthFormat::D24S8:
             return 4;
         default:
-            LOG_CRITICAL(HW_GPU, "Unknown depth format %u", format);
+            LOG_CRITICAL(HW_GPU, "Unknown depth format %u", static_cast<u32>(format));
             UNIMPLEMENTED();
         }
     }
@@ -270,7 +270,7 @@ struct FramebufferRegs {
         case DepthFormat::D24S8:
             return 24;
         default:
-            LOG_CRITICAL(HW_GPU, "Unknown depth format %u", format);
+            LOG_CRITICAL(HW_GPU, "Unknown depth format %u", static_cast<u32>(format));
             UNIMPLEMENTED();
         }
     }
